requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.0.3
celery==5.3.6
APScheduler==3.10.4
beautifulsoup4==4.12.3
requests==2.31.0
selenium==4.20.0
rapidfuzz==3.8.1
gunicorn==22.0.0
pytest==8.1.1
pytest-flask==1.3.0
pytest-mock==3.14.0
